[FEAT] add Magento 2.4.x support
magento is not really consistent with semantic versioning, so `^103.0.1`
is actually a bit brave assumption, but according to our tests extension
still works with Magento 2.4.0 and Magento 2.4.1 which was released
recently and php 7.4.

Leaving compatibility with Magento 2.3.x for now, cause I'm sure there
are some shops that will be still running 2.3.x for a while and 2.3.5
still receives security updates.

see

https://magento.com/sites/default/files/magento-software-lifecycle-policy.pdf